Implement simple .OBJ file parser
This just parses the few things we need/want from the .OBJ file (triangle soup,
basically), and ignores everything else.

Yes this uses Boost.Spirit and no you don't really need it for a file format
this simple, but it's just the least ugly way to implement such a thing quickly.