Implement BVH build & traversal logic
This commit implements a first working version of a BVH data structure
for accelerated ray-triangle intersection.

The way this is currently implemented really is really to 'just make
it work', the code is completely unoptimized, BVH building uses the
most naive splitting strategy, etc. The challenge now is to see how
far we can optimize all this.