Implement front-to-back BVH traversal
While traversing the BVH, child nodes to recurse into are now processed in
front-to-back order.

Unfortunately for the bunny test, for the camera view we have been using so far
this does not improve render performance, in fact it increases render time by ~15%.
This can be explained by the fact that the bunny model has very little 'overdraw'
no matter which order you traverse it in, it is almost a convex model, and except
for rays hitting its ears pretty much all rays will hit the model exactly once.

The front-to-back traversal does have a very large effect on render time of models
that would have significant overdraw when rendered with a traditional back-to-front
rasterizer (painters algorithm). For example rendering the rear-view camera point for
the house scene takes exactly the same time as the front-view camera point when
using ordered traversal, while unordered/arbitrary traversal the rear-view camera
point takes twice as long to render. For this reason, we keep the ordered traversal.