Implement surface area heuristic for BVH construction
When splitting BVH nodes, a number of split planes (currently 8) is now distributed
in each XYZ dimension, and a surface area heuristic is calculated based on the
triangle counts and bounding box area at both sides of each candidate plane. The
plane with the best score is chosen.

This significantly improves render time especially for ‘less convex’ models,
like the digger and watch tower models. Bunny render time is a little faster
but not much. The BVH build time is significantly higher now though, but still
in the order of a few hundred ms even for 100K+ triangle models. BVH build time
can still be improved by using a binning strategy, but this is not really a
priority for now.