Implement turntable render animation
Instead of rendering a static PPM, the main executable now opens
a window and draws the model until the window is closed, rotating
the camera to create a turntable effect. In addition to that, an
ImGUI overlay is added to display the frame time and ray throughput.

This allows more interesting performance analysis, e.g. render
performance dependency on camera angle. Also, it just looks pretty
cool :-)

Additional controls may be added in the future to have some knobs
to tweak rendering etc.