Linearize triangle ID's after building BVH
For efficiency and simplicity, the BVH build process does not swap triangles while
subdividing nodes, but instead uses triangle ID's which are swapped instead. This
indirection prevents having to swap the full triangle data multiple times at BVH
build time, but the end result is that there potential reduced data locality of triangles that end up in the same node. To prevent this, triangle ID's are now linearized after the BVH is built, by copying them back in-order, such that
triangles in the same node are always linearized in memory. This gives a small
improvement in render time especially for models with high poly counts.